created a wrapper resource to fix conflits between drug and test ordes in POC
