+++ api/src/main/java/org/openmrs/module/poc/pocheuristic/dao/PocHeuristicCAO.java
@@ -28,5 +28,6 @@ public interface PocHeuristicCAO {
 	
 	List<Encounter> findEncountersWithTestOrdersByPatient(String patientUUID, boolean voided);
 	
-	List<Encounter> findEncountersByPatientAndEncounterType(final Patient patient, final EncounterType encounterType);
+	List<Encounter> findEncountersByPatientAndEncounterTypeAndOrderTypeUuid(final Patient patient,
+	        final EncounterType encounterType, String orderTypeUuid);
 }
